feat: agregar componente para gestionar imágenes de tipos de habitación
